Simplify all region ci pipeline
This removes unsupported regions, as well as the "run all" functionality
(because it would introduce race conditions with the current
implementation). The latter should be temporary, as in the near future
I would like to work towards be able to run all regions simultaneously,
as this would save a considerable amount of time for each pipeline run.
Also, there are updates to the slack webhook configuration. And lastly,
the region jobs have been split out into deploy, test, and delete,
for clearer views of where things failed, as well as more fine grained
control on retrying certain steps.