Switch away from shell=True
Beyond the security concerns that the world seems to raise about using
shell=True in popen, it seems to be eating non-zero return codes. Commands that failed would hand back a return code of 0 through the shell. Because of this change, instead of using bash quoting to mark argument groupings, we just pass the arguments in as a list in all the places its used.